Efficiency Testing
i have put a couple of ways to test the time efficiency of the user,
cpu and memory efficiency of out program. Edwin was going to help me
put all this info into the report file because no file was generated
when i hit the gen report button on the GUI, thus  could not test it.